+++ Task_11/HMT_1/OrderDetails.cs
@@ -10,11 +10,11 @@ namespace HMT_1
     {
         public int ProductID { get; private set; }
         public string ProductName { get; private set; }
-        public int Quantity { get; private set; }
+        public short Quantity { get; private set; }
         public double UnityPrice { get; private set; }
         public double Discont { get; private set; }
 
-        public OrderDetails(int orderID, string customerID, int employeeID, DateTime orderDate, DateTime shippedDate, string adress,int productID,string productName, int quality, double price, double discont): base(orderID, customerID,employeeID, orderDate, shippedDate, adress)
+        public OrderDetails(int orderID, string customerID, int employeeID, DateTime orderDate, DateTime shippedDate, string adress,int productID,string productName, short quality, double price, double discont): base(orderID, customerID,employeeID, orderDate, shippedDate, adress)
         {
             this.ProductID = productID;
             this.ProductName = productName;
